backend/requirements.txt
fastapi
uvicorn[standard]
numpy
pandas
pydantic
python-dotenv
SQLAlchemy
fredapi